parapet-data/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "parapet-data"
version = "0.1.0"
description = "Mirror-based corpus curation for prompt injection classifiers"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0,<3",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
